refactor(middleware): rename adminAuth.ts to AdminAuth.ts and update imports
